% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camp_names.r
\name{camp_names}
\alias{camp_names}
\title{names of camps from unhcr api for a country}
\usage{
camp_names(name_country)
}
\arguments{
\item{name_country}{country name}
}
\value{
dataframe
}
\description{
names of camps from unhcr api for a country
}
\examples{
camp_names('burundi')

}
