% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camp_popn.r
\name{camp_popn}
\alias{camp_popn}
\title{camp population over time from unhcr api for a camp specified by geo_id
e.g. from :
https://data2.unhcr.org/api/population/get/timeseries?geo_id=933&frequency=day&population_collection=10}
\usage{
camp_popn(geo_id = NULL, name_camp = NULL)
}
\arguments{
\item{geo_id}{id for the camp}

\item{name_camp}{camp name}
}
\value{
dataframe
}
\description{
camp population over time from unhcr api for a camp specified by geo_id
e.g. from :
https://data2.unhcr.org/api/population/get/timeseries?geo_id=933&frequency=day&population_collection=10
}
\examples{
camp_popn(name_camp="Maniema")
camp_popn(933)
camp_popn(499)

}
