% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_id.r
\name{geo_id}
\alias{geo_id}
\title{geo_id for a camp from the name from unhcr api}
\usage{
geo_id(name_camp)
}
\arguments{
\item{name_camp}{name of the camp}
}
\value{
string
}
\description{
geo_id for a camp from the name from unhcr api
}
\examples{
geo_id("Katanga")

}
